+++ src/my_test.rb
@@ -1,12 +1,13 @@
-require('./TestItem')
-require('./NormalRange')
+Dir.glob(File.join('.', '**', '*.rb')).each do |file|
+  require file
+end
 
-test_item = TestItem.new('H1Hblc', :male)
-test_item << NormalRange.new(nil, 4.5, 'B')
-test_item << NormalRange.new(4.6, 6.1, 'A')
-test_item << NormalRange.new(6.2, 6.7, 'B')
-test_item << NormalRange.new(6.8, 7.1, 'C')
-test_item << NormalRange.new(7.2, nil, 'D')
+test_item = HealthDiagnotic::TestItem.new('H1Hblc', :male)
+test_item << HealthDiagnotic::NormalRange.new(nil, 4.5, 'B')
+test_item << HealthDiagnotic::NormalRange.new(4.6, 6.1, 'A')
+test_item << HealthDiagnotic::NormalRange.new(6.2, 6.7, 'B')
+test_item << HealthDiagnotic::NormalRange.new(6.8, 7.1, 'C')
+test_item << HealthDiagnotic::NormalRange.new(7.2, nil, 'D')
 
 puts test_item.determine_result_code(3)
 puts test_item.determine_result_code(4.5)
@@ -23,5 +24,3 @@ puts test_item.determine_result_code(7.2)
 puts test_item.determine_result_code(10)
 puts test_item.be_under_treatment?
 p test_item.normal_range
-
-test_item2 = TestItem.new('H1Hblc', :males)
